packages/scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "YouTube channel scraper using yt-dlp"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp>=2025.12.8",
    "supabase>=2.10.0",
    "python-dotenv>=1.0.0",
    "youtube-transcript-api>=1.2.3",
    "anthropic>=0.76.0",
    "openai>=1.0.0",
    "httpx>=0.28.0",
]

[project.scripts]
scrape-channel = "scraper.channel:main"
scrape-to-db = "scraper.scrape_to_db:main"
summarize = "scraper.summarize:main"
extract-people = "scraper.extract_people:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper"]
